+++ src/Twilio/Rest/FlexApi/V1/ConfigurationInstance.php
@@ -60,6 +60,8 @@ use Twilio\Version;
  * @property bool $flexInsightsDrilldown
  * @property string $flexUrl
  * @property array[] $channelConfigs
+ * @property array $debuggerIntegration
+ * @property array $flexUiStatusReport
  */
 class ConfigurationInstance extends InstanceResource {
     /**
@@ -116,6 +118,8 @@ class ConfigurationInstance extends InstanceResource {
             'flexInsightsDrilldown' => Values::array_get($payload, 'flex_insights_drilldown'),
             'flexUrl' => Values::array_get($payload, 'flex_url'),
             'channelConfigs' => Values::array_get($payload, 'channel_configs'),
+            'debuggerIntegration' => Values::array_get($payload, 'debugger_integration'),
+            'flexUiStatusReport' => Values::array_get($payload, 'flex_ui_status_report'),
         ];
 
         $this->solution = [];