+++ src/Twilio/Rest/Verify/V2.php
@@ -13,6 +13,7 @@ use Twilio\Domain;
 use Twilio\Exceptions\TwilioException;
 use Twilio\InstanceContext;
 use Twilio\Rest\Verify\V2\FormList;
+use Twilio\Rest\Verify\V2\SafelistList;
 use Twilio\Rest\Verify\V2\ServiceList;
 use Twilio\Rest\Verify\V2\TemplateList;
 use Twilio\Rest\Verify\V2\VerificationAttemptList;
@@ -21,16 +22,19 @@ use Twilio\Version;
 
 /**
  * @property FormList $forms
+ * @property SafelistList $safelist
  * @property ServiceList $services
  * @property VerificationAttemptList $verificationAttempts
  * @property VerificationAttemptsSummaryList $verificationAttemptsSummary
  * @property TemplateList $templates
  * @method \Twilio\Rest\Verify\V2\FormContext forms(string $formType)
+ * @method \Twilio\Rest\Verify\V2\SafelistContext safelist(string $phoneNumber)
  * @method \Twilio\Rest\Verify\V2\ServiceContext services(string $sid)
  * @method \Twilio\Rest\Verify\V2\VerificationAttemptContext verificationAttempts(string $sid)
  */
 class V2 extends Version {
     protected $_forms;
+    protected $_safelist;
     protected $_services;
     protected $_verificationAttempts;
     protected $_verificationAttemptsSummary;
@@ -51,6 +55,13 @@ class V2 extends Version {
             $this->_forms = new FormList($this);
         }
         return $this->_forms;
+    }
+
+    protected function getSafelist(): SafelistList {
+        if (!$this->_safelist) {
+            $this->_safelist = new SafelistList($this);
+        }
+        return $this->_safelist;
     }
 
     protected function getServices(): ServiceList {